+++ example/lib/main.dart
@@ -3,8 +3,7 @@ import 'package:flutter/material.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await DartFusion.initialize(
-      localizationAPI: '4Z5H0ZS-QHZM2Z8-NTYP640-38D9RFF');
+  await DartFusion.initialize();
   runApp(const MyApp());
 }